KIALI-2580 Splitting Filters and Sorts
