+++ src/pages/Overview/Filters.ts
@@ -1,10 +1,9 @@
 import { ActiveFilter, FILTER_ACTION_APPEND, FilterTypeWithFilter, FilterValue } from '../../types/Filters';
 import { DEGRADED, FAILURE, HEALTHY } from '../../types/Health';
 import { NamespaceInfo } from './NamespaceInfo';
-import { SortField } from '../../types/SortFilters';
 import { MTLSStatuses } from '../../types/TLSStatus';
 
-export namespace FiltersAndSorts {
+export namespace Filters {
   export const nameFilter: FilterTypeWithFilter<NamespaceInfo> = {
     id: 'namespace_search',
     title: 'Name',
@@ -18,18 +17,9 @@ export namespace FiltersAndSorts {
   };
 
   export const mtlsValues: FilterValue[] = [
-    {
-      id: 'enabled',
-      title: 'Enabled'
-    },
-    {
-      id: 'partiallyEnabled',
-      title: 'Partially Enabled'
-    },
-    {
-      id: 'notEnabled',
-      title: 'Not Enabled'
-    }
+    { id: 'enabled', title: 'Enabled' },
+    { id: 'partiallyEnabled', title: 'Partially Enabled' },
+    { id: 'notEnabled', title: 'Not Enabled' }
   ];
 
   const statusMap = new Map<string, string>([
@@ -54,18 +44,9 @@ export namespace FiltersAndSorts {
   };
 
   const healthValues: FilterValue[] = [
-    {
-      id: FAILURE.name,
-      title: FAILURE.name
-    },
-    {
-      id: DEGRADED.name,
-      title: DEGRADED.name
-    },
-    {
-      id: HEALTHY.name,
-      title: HEALTHY.name
-    }
+    { id: FAILURE.name, title: FAILURE.name },
+    { id: DEGRADED.name, title: DEGRADED.name },
+    { id: HEALTHY.name, title: HEALTHY.name }
   ];
 
   const summarizeHealthFilters = (healthFilters: ActiveFilter[]) => {
@@ -136,65 +117,5 @@ export namespace FiltersAndSorts {
     });
 
     return filteredNamespaces;
-  };
-
-  export const sortFields: SortField<NamespaceInfo>[] = [
-    {
-      id: 'namespace',
-      title: 'Name',
-      isNumeric: false,
-      param: 'ns',
-      compare: (a: NamespaceInfo, b: NamespaceInfo) => a.name.localeCompare(b.name)
-    },
-    {
-      id: 'health',
-      title: 'Status',
-      isNumeric: false,
-      param: 'h',
-      compare: (a: NamespaceInfo, b: NamespaceInfo) => {
-        if (a.status && b.status) {
-          let diff = b.status.inError.length - a.status.inError.length;
-          if (diff !== 0) {
-            return diff;
-          }
-          diff = b.status.inWarning.length - a.status.inWarning.length;
-          if (diff !== 0) {
-            return diff;
-          }
-        } else if (a.status) {
-          return -1;
-        } else if (b.status) {
-          return 1;
-        }
-        // default comparison fallback
-        return a.name.localeCompare(b.name);
-      }
-    },
-    {
-      id: 'mtls',
-      title: 'mTLS',
-      isNumeric: false,
-      param: 'm',
-      compare: (a: NamespaceInfo, b: NamespaceInfo) => {
-        if (a.tlsStatus && b.tlsStatus) {
-          return a.tlsStatus.status.localeCompare(b.tlsStatus.status);
-        } else if (a.tlsStatus) {
-          return -1;
-        } else if (b.tlsStatus) {
-          return 1;
-        }
-
-        // default comparison fallback
-        return a.name.localeCompare(b.name);
-      }
-    }
-  ];
-
-  export const sortFunc = (
-    allNamespaces: NamespaceInfo[],
-    sortField: SortField<NamespaceInfo>,
-    isAscending: boolean
-  ) => {
-    return allNamespaces.sort(isAscending ? sortField.compare : (a, b) => sortField.compare(b, a));
   };
 }
